requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
executing==2.2.0
fastapi==0.115.7
filelock==3.17.0
Flask==3.1.0
fonttools==4.55.5
gdown==5.2.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
logger==1.4
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
pure_eval==0.2.3
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.45.3
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
waitress==3.0.2
wcwidth==0.2.13
Werkzeug==3.1.3
